% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\docType{package}
\name{lab4-package}
\alias{lab4}
\alias{lab4-package}
\title{Lab4 Package}
\description{
This package provides tools for working with linear regression models, including
  a `linreg` class for building and analyzing linear regression models. It includes comprehensive 
  documentation and examples to help you get started.
}
\details{
Lab4

The following package contains a class for a linear regression model called "linreg". 
To get more info about the linear regression model class, run "?linreg".
You can also access coding examples in the vignette by running "vignette(package = "lab4")".
}
